refactor: Break down Medicine Scheduler into multiple components
- Split Medicine Scheduler into smaller, manageable components for better organization.
- Introduced context for centralized state management across the application.
- Updated Medicine model to store time slots in the database.
- Added alert functionality for the Add Medicine modal to improve user feedback.
